Refactored peer calndar seperating concerns
